GithubCI: switch to MacOS 13
Change MacOS version used in CI to 13, as version 12 got
deprecated [1] and causes the jobs to be cancelled.

Install specific versions of Xamarin.iOS and Xamarin.Android
(the same as on MacOS 12), because with default versions the
build fails with error:
```
/Users/runner/work/geewallet/geewallet/src/GWallet.Frontend.XF.iOS/GWallet.Frontend.XF.iOS.fsproj(92,11): error MSB4226: The imported project "/Library/Frameworks/Mono.framework/Versions/6.12.0/lib/mono/xbuild/Xamarin/iOS/Xamarin.iOS.FSharp.targets" was not found. Also, tried to find "Xamarin/iOS/Xamarin.iOS.FSharp.targets" in the fallback search path(s) for $(MSBuildExtensionsPath) - "/Library/Frameworks/Mono.framework/External/xbuild/" . These search paths are defined in "/Library/Frameworks/Mono.framework/Versions/6.12.0/lib/mono/msbuild/Current/bin/MSBuild.dll.config". Confirm that the path in the <Import> declaration is correct, and that the file exists on disk in one of the search paths.
```

One thing to note is that now `macOS--mono-only` and
`macOS--dotnet6-only` take more time (20-25 minutes) and
became the slowest jobs.

[1] https://github.com/actions/runner-images/issues/10721